사라진 파일들 다시 받아서 백업

+++ moneytree_back/src/main/java/com/moneytree_back/domain/DepositRate.java
@@ -5,6 +5,7 @@ import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -26,10 +27,10 @@ public class DepositRate {
     private DepositProduct depositProduct; // 외래 키
 
     @Column(name = "deposit_interest_rate", nullable = false, precision = 5, scale = 2)
-    private Double depositInterestRate;
+    private BigDecimal depositInterestRate;
 
     @Column(name = "deposit_special_rate", precision = 5, scale = 2)
-    private Double depositSpecialRate; // 사용 여부는 선택
+    private BigDecimal depositSpecialRate; // 사용 여부는 선택
 
     @CreationTimestamp
     @Column(name = "deposit_rate_created_at", updatable = false)
